rename calc_rdm module into calc and adjusted test of euclidean distance rdm
