+++ pyrsa/rdm/__init__.py
@@ -1,6 +1,6 @@
 from pyrsa.rdm.rdms import RDMs
-from pyrsa.rdm.calc_rdm import calc_rdm
-from pyrsa.rdm.calc_rdm import calc_rdm_euclid
-from pyrsa.rdm.calc_rdm import calc_rdm_mahalanobis
-from pyrsa.rdm.calc_rdm import calc_rdm_crossnobis
-from pyrsa.rdm.calc_rdm import calc_rdm_correlation
+from pyrsa.rdm.calc import calc_rdm
+from pyrsa.rdm.calc import calc_rdm_euclid
+from pyrsa.rdm.calc import calc_rdm_mahalanobis
+from pyrsa.rdm.calc import calc_rdm_crossnobis
+from pyrsa.rdm.calc import calc_rdm_correlation
