+++ pyrsa/rdm/calc.py
